Rename DetailInformation to ToolTip. Fixed location and name of extensions. Remove unnecessary comments.
