+++ sources/Domain/DataModel/ValueContainers/UVContainer.cs
@@ -1,11 +1,12 @@
 ﻿using Autodesk.Revit.DB;
 using RevitDBExplorer.Domain.DataModel.ValueContainers.Base;
+using RevitDBExplorer.Extensions.System;
 
 // (c) Revit Database Explorer https://github.com/NeVeSpl/RevitDBExplorer/blob/main/license.md
 
 namespace RevitDBExplorer.Domain.DataModel.ValueContainers
 {
-    internal sealed class UVContainer : Base.ValueContainer<UV>, IHaveDetailInformation
+    internal sealed class UVContainer : Base.ValueContainer<UV>, IHaveToolTip
     {
         protected override bool CanBeSnoooped(UV uv) => false;
         protected override string ToLabel(UV uv)
@@ -13,10 +14,7 @@ namespace RevitDBExplorer.Domain.DataModel.ValueContainers
             return $"({uv.U}, {uv.V})";
         }
 
-        /// <summary>
-        /// Retrieve uv in mm
-        /// </summary>
-        public string DetailInformationText
+        public string ToolTip
         {
             get => $"{ToLabel(Value)}\n({Value.U.ToLengthDisplayString(Units)}, {Value.V.ToLengthDisplayString(Units)})";
         }
